+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * [dropwizard]
   - Fix dropwizard 1.x compatibility
   - Make dropwizard dependency "provided" (to avoid version clashes)
+  - Rename `ConfigUpdaterBundle` into `UpdateConfigBundle` to unify naming with command
 
 ### 1.0.0 (2021-08-29)
 * Initial release